[WIP] Fixing errors
