+++ vue/config/local.env.js
@@ -1,4 +1,4 @@
 exports.token = {
-  github: '',
+  github: '67e29c3341d8b95ed00759272ae80043a8e22c18',
   bitbucket: ''
 }
