+++ vue/src/views/InstanceDetail.vue
@@ -1,0 +1,19 @@
+<template>
+    <div class="row">
+    <p>Instance Detail Page</p>
+
+    <h1>{{this.$route.params.user}}</h1>
+
+    <h1>{{this.$route.params.repo}}</h1>
+
+    </div>
+</template>
+
+<script>
+
+</script>
+
+
+<style scoped>
+
+</style>
